move indexed collection out
