+++ core/incremental/build.ts
@@ -11,7 +11,7 @@ import {
 } from "../../util/util";
 import { ppb } from "../pretty";
 import { List } from "immutable";
-import { emptyIndexedCollection } from "./indexedCollection";
+import { emptyIndexedCollection } from "../../util/indexedCollection";
 import { fastPPT } from "./fastPPT";
 
 export function declareTable(graph: RuleGraph, name: string): RuleGraph {
