+++ util/indexedCollection.ts
